% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatio.R
\docType{package}
\name{rmatio}
\alias{rmatio}
\alias{rmatio-package}
\title{\pkg{rmatio}: reading and writing Matlab MAT files from R}
\description{
Reading and writing Matlab MAT files from R
}
\details{
\code{rmatio} supports reading MAT version 4, MAT version 5 and
MAT compressed version 5.

\code{rmatio} can write version 5 MAT files and version 5 files
with variable compression.
}
\references{
\itemize{
  \item Christopher C. Hulbert, MATIO User Manual for version 1.5.2.\cr
  \url{https://sourceforge.net/projects/matio/files/matio/1.5.2/matio_user_guide.pdf}

  \item The MathWorks Inc., MATLAB - MAT-File Format, version R2013b, September 2013.\cr
  \url{https://www.mathworks.com/help/pdf_doc/matlab/matfile_format.pdf}
}
}
